# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.16.0

package db

import (
	"time"
)

type Address struct {
	ID        int64     `json:"id"`
	Address   string    `json:"address"`
	Username  string    `json:"username"`
	District  string    `json:"district"`
	City      string    `json:"city"`
	CreatedAt time.Time `json:"created_at"`
}

type Book struct {
	ID          int64     `json:"id"`
	Name        string    `json:"name"`
	Price       float64   `json:"price"`
	Image       []string  `json:"image"`
	Description string    `json:"description"`
	Author      string    `json:"author"`
	Publisher   string    `json:"publisher"`
	Quantity    int32     `json:"quantity"`
	CreatedAt   time.Time `json:"created_at"`
}

type BooksGenre struct {
	ID        int64     `json:"id"`
	BooksID   int64     `json:"books_id"`
	GenresID  int64     `json:"genres_id"`
	CreatedAt time.Time `json:"created_at"`
}

type BooksSubgenre struct {
	ID          int64     `json:"id"`
	BooksID     int64     `json:"books_id"`
	SubgenresID int64     `json:"subgenres_id"`
	CreatedAt   time.Time `json:"created_at"`
}

type Cart struct {
	ID        int64     `json:"id"`
	BooksID   int64     `json:"books_id"`
	Username  string    `json:"username"`
	CreatedAt time.Time `json:"created_at"`
}

type Genre struct {
	ID        int64     `json:"id"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
}

type Order struct {
	ID        int64     `json:"id"`
	Username  string    `json:"username"`
	CreatedAt time.Time `json:"created_at"`
}

type Review struct {
	ID        int64     `json:"id"`
	Username  string    `json:"username"`
	BooksID   int64     `json:"books_id"`
	Comments  string    `json:"comments"`
	Rating    int32     `json:"rating"`
	CreatedAt time.Time `json:"created_at"`
}

type Subgenre struct {
	ID        int64     `json:"id"`
	GenresID  int64     `json:"genres_id"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
}

type Transaction struct {
	ID        int64     `json:"id"`
	OrdersID  int64     `json:"orders_id"`
	BooksID   int64     `json:"books_id"`
	CreatedAt time.Time `json:"created_at"`
}

type User struct {
	Username    string    `json:"username"`
	FullName    string    `json:"full_name"`
	Email       string    `json:"email"`
	Password    string    `json:"password"`
	Image       string    `json:"image"`
	PhoneNumber string    `json:"phone_number"`
	Role        string    `json:"role"`
	CreatedAt   time.Time `json:"created_at"`
}

type Wishlist struct {
	ID        int64     `json:"id"`
	BooksID   int64     `json:"books_id"`
	Username  string    `json:"username"`
	CreatedAt time.Time `json:"created_at"`
}